finish beta

+++ HW_07/phonebook/logic.py
@@ -1,6 +1,8 @@
+from ast import Import
 import json
 from importlib.resources import path
 import os
+import view
 
 
 FILE_NAME = 'contacts_book.json'
@@ -13,20 +15,87 @@ def add_contact ():
     comment = input('Enter a comment: ')
     abonent = {'first_name': first_name, 'last_name': last_name, 'phone_number': phone_num, 'comment' : comment}
     with open (path, "r") as file:
+        if abonents == "None":
+            abonents = []
         abonents: list = json.load(file)
     with open(path, "w") as file:
         abonents.append(abonent)
-        json.dump(abonents, file)
-    #     я закоментил сохранение в txt и csv потому что смысла нет, лучше хранить в одном месте
-    #     а если надо будет выгрузить в другой формат, то лучше это делать через функцию.
-    # with open ('HW_07' + os.sep + 'phonebook' + os.sep + 'phonebook.txt', "a") as file:
-    #     file.write(f'{abonent}\n')
-    # with open ('HW_07' + os.sep + 'phonebook' + os.sep + 'phonebook.csv', "a") as file:
-    #     file.write(f'{abonent}\n')
+        json.dump(abonents, file, delimeter=3)
 
-
-def find_contact(first_name: str, last_name: str):
+def find_contact():
     with open(path, 'r') as file:
         abonents: list[dict] = json.load(file)
-        # TODO: реализовать поиск нужного контакта в списке контактов abonents
-        return
+        dict_count = 0
+        # while dict_count < 1: TODO add while search
+        find_ab = input('Enter the name of the abonent to find: ')
+        for dicts in abonents:
+            if find_ab.lower() in dicts['first_name'].lower() + ' ' + dicts['last_name'].lower():
+                view.print_dict(dicts)
+                dict_count += 1
+        if dict_count == 0:
+            print('There\'s no contact with this name. Try again')
+
+def delite_contact():
+    with open(path, 'r') as file:
+        abonents: list[dict] = json.load(file)
+        dict_count = 0
+        find_ab = input('Enter the name of the abonent to delite: ')
+        for dicts in abonents:
+            if find_ab.lower() in dicts['first_name'].lower() + ' ' + dicts['last_name'].lower():
+                dict_count += 1
+                print(f'Are you sure you want to remove contact: {dicts["first_name"]} {dicts["last_name"]}?')
+                del_confirmation = input('for yes - press 1\nfor no - press 2:\n')
+                if del_confirmation == '1':
+                    abonents.remove(dicts)
+                    with open(path, "w") as file:
+                        json.dump(abonents, file) # TODO why id delite item 1 quite?
+                    view.success_update()
+        if dict_count == 0:
+            print('There\'s no contact with this name. Try again')
+
+def change_contact():
+    with open(path, 'r') as file:
+        abonents: list[dict] = json.load(file)
+        dict_count = 0
+        find_ab = input("Enter the name of the abonent to change it's data: ")
+        for dicts in abonents:
+            if find_ab.lower() in dicts['first_name'].lower() + ' ' + dicts['last_name'].lower():
+                dict_count += 1
+                print(f'Are you sure you want to change data for contact: {dicts["first_name"]} {dicts["last_name"]}?')
+                change_confirmation = input('for yes - press 1\nfor no - press 2:\n')
+                if change_confirmation == '1':
+                    change_data = input(f'To change the data enter the number\n1 - first_name\n2 - last_name\n3 - phone_number\n4 - comment\n:')
+                    match change_data:
+                        case '1': 
+                            new_name = input('Enter a new first_name: ')
+                            dicts["first_name"] = new_name
+                        case '2': 
+                            new_name = input('Enter a new last_name: ')
+                            dicts["last_name"] = new_name
+                        case '3': 
+                            new_name = input('Enter a new phone_number: ')
+                            dicts["phone_number"] = new_name
+                        case '4': 
+                            new_name = input('Enter a new comment: ')
+                            dicts["comment"] = new_name
+                    with open(path, "w") as file:
+                        json.dump(abonents, file)
+                    view.success_update()    
+        if dict_count == 0:
+            print('There\'s no contact with this name. Try again')
+                        
+                        
+
+
+
+
+
+
+# d = [{"first_name": "Maria", "last_name": "Aksenova", "phone_number": "89078909887", "comment": "no"}, {"first_name": "Anya", "last_name": "Rihter", "phone_number": "89055567788", "comment": "friend"}]
+# # find_ab = 'Mar'
+# for dicts in d:
+# #     if find_ab in dicts['first_name'] + ' ' + dicts['last_name']: # TODO add lower search
+# #         print('pop')
+# #         d.remove(dicts)
+
+#     print(dicts["first_name"], dicts['last_name'])
